requirements.txt
flashtext==2.7
Jinja2==3.1.2
nltk==3.8.1
numpy==1.25.1
openai==1.34.0
pandas==2.0.3
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.1
ReFinED==1.0
Requests==2.32.3
ruamel.base==1.0.0
scikit_learn==1.3.0
sentence_transformers==2.6.1
tenacity==8.2.2
torch==2.1.0
tqdm==4.65.0
transformers==4.40.2
wikidataintegrator==0.9.30
